requirements.txt
python-telegram-bot==20.7
ccxt
requests
asyncio
python-binance
bybit
